Updated command bar to have search suggestions from Google and arrow keys. It still doesn't work for updating urls with Command+L. Also added code to change colors for icons and text in theme.
