+++ iPad browser/Helpers/Suggestions/SuggestionsData.swift
@@ -178,398 +178,140 @@ struct SuggestionsView: View {
 
 
 
-let suggestionUrls = [
-  "google.com",
-  "apple.com",
-  "youtube.com",
-  "wikipedia.org",
-  "amazon.com",
-  "github.com",
-  "nytimes.com",
-  "notion.so",
-  "linkedin.com",
-  "wordpress.org",
-  "support.google.com",
-  "play.google.com",
-  "microsoft.com",
-  "docs.google.com",
-  "maps.google.com",
-  "en.wikipedia.org",
-  "cloudflare.com",
-  "whatsapp.com",
-  "youtu.be",
-  "plus.google.com",
-  "accounts.google.com",
-  "bp.blogspot.com",
-  "blogspot.com",
-  "mozilla.org",
-  "drive.google.com",
-  "europa.eu",
-  "sites.google.com",
-  "adobe.com",
-  "googleusercontent.com",
-  "facebook.com",
-  "www.blogger.com",
-  "istockphoto.com",
-  "t.me",
-  "uol.com.br",
-  "policies.google.com",
-  "vimeo.com",
-  "gravatar.com",
-  "search.google.com",
-  "forbes.com",
-  "dropbox.com",
-  "myspace.com",
-  "cnn.com",
-  "medium.com",
-  "gstatic.com",
-  "google.com.br",
-  "slideshare.net",
-  "sheets.google.com",
-  "etsy.com",
-  "dailymotion.com",
-  "google.es",
-  "tiktok.com",
-  "opera.com",
-  "msn.com",
-  "tools.google.com",
-  "wa.me",
-  "draft.blogger.com",
-  "fr.wikipedia.org",
-  "pt.wikipedia.org",
-  "developers.google.com",
-  "nih.gov",
-  "bbc.co.uk",
-  "healthline.com",
-  "brandbucket.com",
-  "macys.com",
-  "realtor.com",
-  "news.google.com",
-  "creativecommons.org",
-  "wikimedia.org",
-  "files.wordpress.com",
-  "google.de",
-  "live.com",
-  "craigslist.org",
-  "imdb.com",
-  "shopify.com",
-  "line.me",
-  "www.yahoo.com",
-  "enable-javascript.com",
-  "es.wikipedia.org",
-  "feedburner.com",
-  "globo.com",
-  "paypal.com",
-  "netvibes.com",
-  "bestbuy.com",
-  "who.int",
-  "theguardian.com",
-  "www.weebly.com",
-  "mail.google.com",
-  "jimdofree.com",
-  "google.co.jp",
-  "afternic.com",
-  "youronlinechoices.com",
-  "google.fr",
-  "elpais.com",
-  "picasaweb.google.com",
-  "fb.com",
-  "t.co",
-  "archive.org",
-  "de.wikipedia.org",
-  "archiveofourown.org",
-  "ru.wikipedia.org",
-  "usps.com",
-  "telegraph.co.uk",
-  "pinterest.com",
-  "mail.ru",
-  "thesun.co.uk",
-  "google.co.uk",
-  "same.energy",
-  "samsung.com",
-  "webmd.com",
-  "weather.com",
-  "weather.gov",
-  "chase.com",
-  "allrecipes.com",
-  "amazon.co.jp",
-  "cpanel.net",
-  "wsj.com",
-  "cvs.com",
-  "get.google.com",
-  "walgreens.com",
-  "independent.co.uk",
-  "indeed.com",
-  "www.gov.uk",
-  "domainmarket.com",
-  "networkadvertising.org",
-  "cdc.gov",
-  "hugedomains.com",
-  "pixabay.com",
-  "twitter.com",
-  "cbsnews.com",
-  "terra.com.br",
-  "abril.com.br",
-  "buydomains.com",
-  "huffingtonpost.com",
-  "nature.com",
-  "booking.com",
-  "w3.org",
-  "reuters.com",
-  "godaddy.com",
-  "ebay.com",
-  "wp.com",
-  "adssettings.google.com",
-  "marketingplatform.google.com",
-  "change.org",
-  "mirror.co.uk",
-  "plesk.com",
-  "namecheap.com",
-  "forms.gle",
-  "dan.com",
-  "bbc.com",
-  "storage.googleapis.com",
-  "scribd.com",
-  "time.com",
-  "usatoday.com",
-  "ig.com.br",
-  "aboutads.info",
-  "issuu.com",
-  "businessinsider.com",
-  "nasa.gov",
-  "4shared.com",
-  "amazon.co.uk",
-  "cnet.com",
-  "office.com",
-  "foxnews.com",
-  "tinyurl.com",
-  "aliexpress.com",
-  "photos.google.com",
-  "amazon.de",
-  "bloomberg.com",
-  "bing.com",
-  "indiatimes.com",
-  "huffpost.com",
-  "telegram.me",
-  "myaccount.google.com",
-  "goo.gl",
-  "bit.ly",
-  "planalto.gov.br",
-  "estadao.com.br",
-  "mediafire.com",
-  "washingtonpost.com",
-  "ytimg.com",
-  "soundcloud.com",
-  "shutterstock.com",
-  "news.yahoo.com",
-  "dailymail.co.uk",
-  "instagram.com",
-  "researchgate.net",
-  "google.it",
-  "fandom.com",
-  "list-manage.com",
-  "t-mobile.com",
-  "un.org",
-  "www.livejournal.com",
-  "calendar.google.com",
-  "deezer.com",
-  "playstation.com",
-  "google.ca",
-  "berkeley.edu",
-  "duckduckgo.com",
-  "ted.com",
-  "it.wikipedia.org",
-  "abcnews.go.com",
-  "smh.com.au",
-  "offset.com",
-  "aol.com",
-  "linktr.ee",
-  "canada.ca",
-  "lemonde.fr",
-  "yadi.sk",
-  "ovh.com",
-  "engadget.com",
-  "psychologytoday.com",
-  "sciencedirect.com",
-  "apache.org",
-  "imageshack.us",
-  "mit.edu",
-  "groups.google.com",
-  "ups.com",
-  "akamaihd.net",
-  "wired.com",
-  "clarin.com",
-  "sky.com",
-  "pbs.org",
-  "yandex.com",
-  "naver.com",
-  "zippyshare.com",
-  "freepik.com",
-  "doubleclick.net",
-  "statista.com",
-  "espn.com",
-  "alicdn.com",
-  "20minutos.es",
-  "ouest-france.fr",
-  "usablenet.com",
-  "whitehouse.gov",
-  "instructables.com",
-  "gooyaabitemplates.com",
-  "translate.google.com",
-  "abc.es",
-  "express.co.uk",
-  "alexa.com",
-  "ft.com",
-  "elmundo.es",
-  "marca.com",
-  "cpanel.com",
-  "gofundme.com",
-  "cornell.edu",
-  "pinterest.fr",
-  "google.pl",
-  "discord.gg",
-  "timeweb.ru",
-  "finance.yahoo.com",
-  "e-monsite.com",
-  "akamaized.net",
-  "vice.com",
-  "redbull.com",
-  "unesco.org",
-  "rapidshare.com",
-  "prezi.com",
-  "000webhost.com",
-  "sapo.pt",
-  "nikkei.com",
-  "springer.com",
-  "netlify.app",
-  "search.yahoo.com",
-  "cambridge.org",
-  "bp2.blogger.com",
-  "ea.com",
-  "taringa.net",
-  "lavanguardia.com",
-  "ikea.com",
-  "qq.com",
-  "id.wikipedia.org",
-  "zillow.com",
-  "mozilla.com",
-  "amazon.fr",
-  "nypost.com",
-  "weibo.com",
-  "www.canalblog.com",
-  "photobucket.com",
-  "yelp.com",
-  "gizmodo.com",
-  "quora.com",
-  "sakura.ne.jp",
-  "unsplash.com",
-  "detik.com",
-  "android.com",
-  "code.google.com",
-  "addtoany.com",
-  "pexels.com",
-  "hindustantimes.com",
-  "nbcnews.com",
-  "goodreads.com",
-  "wellsfargo.com",
-  "wikia.com",
-  "dw.com",
-  "yandex.ru",
-  "theatlantic.com",
-  "windows.net",
-  "ggpht.com",
-  "privacyshield.gov",
-  "discord.com",
-  "doi.org",
-  "www.wix.com",
-  "www.gov.br",
-  "leparisien.fr",
-  "books.google.com",
-  "nydailynews.com",
-  "ipv4.google.com",
-  "mashable.com",
-  "rtve.es",
-  "insider.com",
-  "sciencedaily.com",
-  "bloglovin.com",
-  "urbandictionary.com",
-  "rakuten.co.jp",
-  "hulu.com",
-  "gnu.org",
-  "plos.org",
-  "hubspot.com",
-  "spiegel.de",
-  "google.nl",
-  "cnil.fr",
-  "sedo.com",
-  "ziddu.com",
-  "steampowered.com",
-  "capitalone.com",
-  "costco.com",
-  "rt.com",
-  "icann.org",
-  "kickstarter.com",
-  "francetvinfo.fr",
-  "npr.org",
-  "cointernet.com.co",
-  "nationalgeographic.com",
-  "latimes.com",
-  "outlook.com",
-  "spotify.com",
-  "ria.ru",
-  "cbc.ca",
-  "guardian.co.uk",
-  "cnbc.com",
-  "ovhcloud.com",
-  "tripadvisor.com",
-  "nginx.com",
-  "oracle.com",
-  "yahoo.co.jp",
-  "britannica.com",
-  "huawei.com",
-  "buzzfeed.com",
-  "techcrunch.com",
-  "zoom.us",
-  "wiley.com",
-  "home.pl",
-  "homedepot.com",
-  "standard.co.uk",
-  "photos1.blogger.com",
-  "newyorker.com",
-  "secureserver.net",
-  "hp.com",
-  "ietf.org",
-  "sendspace.com",
-  "amazon.es",
-  "oup.com",
-  "tmz.com",
-  "m.wikipedia.org",
-  "bandcamp.com",
-  "economist.com",
-  "ssl-images-amazon.com",
-  "target.com",
-  "webnode.page",
-  "walmart.com",
-  "surveymonkey.com",
-  "ovh.net",
-  "abc.net.au",
-  "newsweek.com",
-  "zendesk.com",
-  "nginx.org",
-  "dreamstime.com",
-  "ca.gov",
-  "ja.wikipedia.org",
-  "clickbank.net",
-  "nextdoor.com",
-  "bankofamerica.com",
-  "hollywoodreporter.com",
-  "academia.edu",
-  "biglobe.ne.jp",
-  "liveinternet.ru",
-  "sfgate.com",
-  "roblox.com",
-  "google.ru",
-  "gmail.com",
-  "ftc.gov",
-  "netflix.com"
+
+// Sample dataset of search queries
+// Sample dataset of search queries
+let sampleQueries = [
+    "machine learning",
+    "swift programming",
+    "data science",
+    "python tutorial",
+    "deep learning",
+    "web development",
+    "computer science",
+    "java programming",
+    "artificial intelligence",
+    "software engineering",
+    "javascript tutorial",
+    "data analysis",
+    "android development",
+    "iOS app development",
+    "cloud computing",
+    "network security",
+    "computer vision",
+    "natural language processing",
+    "database management",
+    "algorithm design",
+    "mobile application development",
+    "frontend development",
+    "backend development",
+    "game development",
+    "user interface design",
+    "graphic design",
+    "digital marketing",
+    "search engine optimization",
+    "social media marketing",
+    "e-commerce",
+    "machine learning algorithms",
+    "data visualization",
+    "cybersecurity",
+    "blockchain technology",
+    "internet of things",
+    "augmented reality",
+    "virtual reality",
+    "big data analytics",
+    "cloud storage",
+    "computer programming basics",
+    "coding interview preparation",
+    "online courses",
+    "programming languages",
+    "technology trends",
+    "coding bootcamp",
+    "software development tools",
+    "version control systems",
+    "agile methodology",
+    "project management",
+    "productivity tips",
+    "remote work",
+    "work-life balance",
+    "health and wellness",
+    "cooking recipes",
+    "home workouts",
+    "travel destinations",
+    "financial planning",
+    "personal finance",
+    "self-improvement",
+    "mindfulness meditation",
+    "stress management",
+    "relationship advice",
+    "parenting tips",
+    "pet care",
+    "gardening tips",
+    "DIY projects",
+    "home decor ideas",
+    "fashion trends",
+    "beauty tips",
+    "book recommendations",
+    "movie reviews",
+    "music playlists",
+    "video game reviews",
+    "sports news",
+    "celebrity gossip",
+    "political news",
+    "current events"
 ]
+
+
+// Function to preprocess the search queries
+func preprocess(_ query: String) -> [String] {
+    return query.lowercased().components(separatedBy: " ")
+}
+
+// Function to train a Naive Bayes classifier
+func trainModel(_ queries: [String]) -> [String: Int] {
+    var wordCounts: [String: Int] = [:]
+    
+    for query in queries {
+        let tokens = preprocess(query)
+        for token in tokens {
+            wordCounts[token, default: 0] += 1
+        }
+    }
+    
+    return wordCounts
+}
+
+// Function to predict autocomplete suggestions with multiple words
+func predict(_ prefix: String, model: [String: Int]) -> [String] {
+    let prefixTokens = preprocess(prefix)
+    var suggestions: [String] = []
+    
+    // Try to find suggestions starting from each token in the prefix
+    for i in 0..<prefixTokens.count {
+        var currentPrefix = ""
+        for j in i..<prefixTokens.count {
+            currentPrefix += prefixTokens[j] + " "
+            let partialSuggestions = model.keys.filter { $0.hasPrefix(currentPrefix.trimmingCharacters(in: .whitespacesAndNewlines)) }
+            suggestions += partialSuggestions
+        }
+    }
+    
+    // Remove duplicates and return top suggestions
+    suggestions = Array(Set(suggestions).prefix(3))
+    return suggestions
+}
+
+
+// Train the model
+var model = trainModel(suggestionWordsData)
+
+// Function to update model with user input
+func updateModel(with userInput: String) {
+    let tokens = preprocess(userInput)
+    for token in tokens {
+        model[token, default: 0] += 1
+    }
+}
+
+
+
