+++ iPad browser/Views/Main/Sidebar.swift
@@ -22,6 +22,7 @@ struct Sidebar: View {
     @Binding var tabBarShown: Bool
     @Binding var startColor: Color
     @Binding var endColor: Color
+    @Binding var textColor: Color
     @Binding var hoverSpace: String
     var geo: GeometryProxy
     
@@ -65,7 +66,7 @@ struct Sidebar: View {
     
     var body: some View {
         VStack {
-            ToolbarButtonsView(selectedTabLocation: $selectedTabLocation, navigationState: navigationState, pinnedNavigationState: pinnedNavigationState, favoritesNavigationState: favoritesNavigationState, hideSidebar: $hideSidebar, searchInSidebar: $searchInSidebar, commandBarShown: $commandBarShown, tabBarShown: $tabBarShown, startColor: $startColor, endColor: $endColor, geo: geo).frame(height: 40)
+            ToolbarButtonsView(selectedTabLocation: $selectedTabLocation, navigationState: navigationState, pinnedNavigationState: pinnedNavigationState, favoritesNavigationState: favoritesNavigationState, hideSidebar: $hideSidebar, searchInSidebar: $searchInSidebar, commandBarShown: $commandBarShown, tabBarShown: $tabBarShown, startColor: $startColor, endColor: $endColor, textColor: $textColor, geo: geo).frame(height: 40)
             
             // Sidebar Searchbar
             Button {
@@ -141,8 +142,8 @@ struct Sidebar: View {
                 ForEach(favoritesNavigationState.webViews, id:\.self) { tab in
                     ZStack {
                         RoundedRectangle(cornerRadius: 20)
-                            .stroke(Color.white.opacity(tab == favoritesNavigationState.selectedWebView ? 1.0 : hoverTab == tab ? 0.6: 0.2), lineWidth: 3)
-                            .fill(Color(.white).opacity(tab == favoritesNavigationState.selectedWebView ? 0.5 : hoverTab == tab ? 0.15: 0.0001))
+                            .stroke(textColor.opacity(tab == favoritesNavigationState.selectedWebView ? 1.0 : hoverTab == tab ? 0.6: 0.2), lineWidth: 3)
+                            .fill(textColor.opacity(tab == favoritesNavigationState.selectedWebView ? 0.5 : hoverTab == tab ? 0.15: 0.0001))
                             .frame(height: 75)
                         
                         if favoritesStyle {
@@ -272,177 +273,7 @@ struct Sidebar: View {
             // Tabs
             ScrollView {
                 ForEach(pinnedNavigationState.webViews, id: \.self) { tab in
-                    ZStack {
-                        if reloadTitles {
-                            Color.white.opacity(0.0)
-                        }
-                        
-                        RoundedRectangle(cornerRadius: 20)
-                            .foregroundStyle(Color(.white).opacity(tab == pinnedNavigationState.selectedWebView ? 0.5 : hoverTab == tab ? 0.2: 0.0001))
-                            .frame(height: 50)
-                        
-                        HStack {
-                            if faviconLoadingStyle {
-                                WebImage(url: URL(string: "https://www.google.com/s2/favicons?domain=\(tab.url?.absoluteString)&sz=\(128)".replacingOccurrences(of: "https://www.google.com/s2/favicons?domain=Optional(", with: "https://www.google.com/s2/favicons?domain=").replacingOccurrences(of: ")&sz=", with: "&sz=").replacingOccurrences(of: "\"", with: ""))) { image in
-                                    image
-                                        .resizable()
-                                        .scaledToFit()
-                                        .frame(width: 25, height: 25)
-                                        .cornerRadius(50)
-                                        .padding(.leading, 5)
-                                    
-                                } placeholder: {
-                                    LoadingAnimations(size: 25, borderWidth: 5.0)
-                                        .padding(.leading, 5)
-                                }
-                                .onSuccess { image, data, cacheType in
-                                    // Success
-                                    // Note: Data exist only when queried from disk cache or network. Use `.queryMemoryData` if you really need data
-                                }
-                                .indicator(.activity) // Activity Indicator
-                                .transition(.fade(duration: 0.5)) // Fade Transition with duration
-                                .scaledToFit()
-                                
-                            } else {
-                                AsyncImage(url: URL(string: "https://www.google.com/s2/favicons?domain=\(tab.url?.absoluteString)&sz=\(128)".replacingOccurrences(of: "https://www.google.com/s2/favicons?domain=Optional(", with: "https://www.google.com/s2/favicons?domain=").replacingOccurrences(of: ")&sz=", with: "&sz=").replacingOccurrences(of: "\"", with: ""))) { image in
-                                    image
-                                        .resizable()
-                                        .scaledToFit()
-                                        .frame(width: 25, height: 25)
-                                        .cornerRadius(50)
-                                        .padding(.leading, 5)
-                                    
-                                } placeholder: {
-                                    LoadingAnimations(size: 25, borderWidth: 5.0)
-                                        .padding(.leading, 5)
-                                }
-                                
-                            }
-                            
-                            if tab.title == "" {
-                                Text(tab.url?.absoluteString ?? "Tab not found.")
-                                    .lineLimit(1)
-                                    .foregroundColor(Color.foregroundColor(forHex: UserDefaults.standard.string(forKey: "startColorHex") ?? "ffffff"))
-                                    .padding(.leading, 5)
-                                    .onReceive(timer) { _ in
-                                        reloadTitles.toggle()
-                                    }
-                            }
-                            else {
-                                Text(tab.title ?? "Tab not found.")
-                                    .lineLimit(1)
-                                    .foregroundColor(Color.foregroundColor(forHex: UserDefaults.standard.string(forKey: "startColorHex") ?? "ffffff"))
-                                    .padding(.leading, 5)
-                                    .onReceive(timer) { _ in
-                                        reloadTitles.toggle()
-                                    }
-                            }
-                            
-                            Spacer()
-                            
-                            Button(action: {
-                                if let index = pinnedNavigationState.webViews.firstIndex(of: tab) {
-                                    pinnedRemoveTab(at: index)
-                                }
-                            }) {
-                                if hoverTab == tab || pinnedNavigationState.selectedWebView == tab {
-                                    ZStack {
-                                        Color(.white)
-                                            .opacity(hoverCloseTab == tab ? 0.3: 0.0)
-                                        Image(systemName: "xmark")
-                                            .resizable()
-                                            .scaledToFit()
-                                            .frame(width: 15, height: 15)
-                                            .foregroundStyle(Color.white)
-                                            .opacity(hoverCloseTab == tab ? 1.0: 0.8)
-                                        
-                                    }.frame(width: 35, height: 35).cornerRadius(7).padding(.trailing, 10)
-                                        .hoverEffect(.lift)
-                                        .onHover(perform: { hovering in
-                                            if hovering {
-                                                hoverCloseTab = tab
-                                            }
-                                            else {
-                                                hoverCloseTab = WKWebView()
-                                            }
-                                        })
-                                    
-                                }
-                            }
-                        }
-                    }
-                    .contextMenu {
-                        Button {
-                            pinnedNavigationState.createNewWebView(withRequest: URLRequest(url: URL(string: formatURL(from: tab.url?.absoluteString ?? ""))!))
-                        } label: {
-                            Label("Duplicate", systemImage: "plus.square.on.square")
-                        }
-                        
-                        Button {
-                            navigationState.webViews.append(tab)
-                            
-                            if let index = pinnedNavigationState.webViews.firstIndex(of: tab) {
-                                pinnedRemoveTab(at: index)
-                            }
-                        } label: {
-                            Label("Unpin", systemImage: "pin.fill")
-                        }
-                        
-                        Button {
-                            favoritesNavigationState.webViews.append(tab)
-                            
-                            if let index = pinnedNavigationState.webViews.firstIndex(of: tab) {
-                                pinnedRemoveTab(at: index)
-                            }
-                        } label: {
-                            Label("Favorite", systemImage: "star")
-                        }
-                        
-                        Button {
-                            if let index = pinnedNavigationState.webViews.firstIndex(of: tab) {
-                                pinnedRemoveTab(at: index)
-                            }
-                        } label: {
-                            Label("Close Tab", systemImage: "xmark")
-                        }
-                        
-                    }
-                    .onAppear() {
-                        DispatchQueue.main.asyncAfter(deadline: .now() + 5) {
-                            hoverTab = WKWebView()
-                        }
-                    }
-                    .onHover(perform: { hovering in
-                        if hovering {
-                            hoverTab = tab
-                        }
-                        else {
-                            hoverTab = WKWebView()
-                        }
-                    })
-                    .onTapGesture {
-                        navigationState.selectedWebView = nil
-                        navigationState.currentURL = nil
-                        
-                        favoritesNavigationState.selectedWebView = nil
-                        favoritesNavigationState.currentURL = nil
-                        
-                        selectedTabLocation = "pinnedTabs"
-                        
-                        Task {
-                            await pinnedNavigationState.selectedWebView = tab
-                            await pinnedNavigationState.currentURL = tab.url
-                        }
-                        
-                        if let unwrappedURL = tab.url {
-                            searchInSidebar = unwrappedURL.absoluteString
-                        }
-                    }
-                    .onDrag {
-                        self.draggedTab = tab
-                        return NSItemProvider()
-                    }
-                    .onDrop(of: [.text], delegate: DropViewDelegate(destinationItem: tab, allTabs: $pinnedNavigationState.webViews, draggedItem: $draggedTab))
+                    PinnedTab(reloadTitles: $reloadTitles, tab: tab, hoverTab: $hoverTab, faviconLoadingStyle: $faviconLoadingStyle, searchInSidebar: $searchInSidebar, hoverCloseTab: $hoverCloseTab, selectedTabLocation: $selectedTabLocation, draggedTab: $draggedTab, navigationState: navigationState, pinnedNavigationState: pinnedNavigationState, favoritesNavigationState: favoritesNavigationState)
                 }
                 
                 HStack {
@@ -457,7 +288,7 @@ struct Sidebar: View {
                                 .resizable()
                                 .scaledToFit()
                                 .frame(width: 20, height: 20)
-                                .foregroundStyle(Color.white)
+                                .foregroundStyle(textColor)
                                 .opacity(spaceIconHover ? 1.0: 0.5)
                             
                         }.frame(width: 40, height: 40).cornerRadius(7)
@@ -474,11 +305,11 @@ struct Sidebar: View {
                     
                     
                     Text(currentSpace)
-                        .foregroundStyle(Color.white)
+                        .foregroundStyle(textColor)
                         .opacity(0.5)
                         .font(.system(.caption, design: .rounded, weight: .medium))
                     
-                    Color.white
+                    textColor
                         .opacity(0.5)
                         .frame(height: 1)
                         .cornerRadius(10)
@@ -510,7 +341,7 @@ struct Sidebar: View {
                             .frame(height: 50)
                         HStack {
                             Label("New Tab", systemImage: "plus")
-                                .foregroundStyle(Color.white)
+                                .foregroundStyle(textColor)
                                 .font(.system(.headline, design: .rounded, weight: .bold))
                                 .padding(.leading, 10)
                             
@@ -528,173 +359,7 @@ struct Sidebar: View {
                 }
                 
                 ForEach(navigationState.webViews.reversed(), id: \.self) { tab in
-                    ZStack {
-                        if reloadTitles {
-                            Color.white.opacity(0.0)
-                        }
-                        
-                        RoundedRectangle(cornerRadius: 20)
-                            .foregroundStyle(Color(.white).opacity(tab == navigationState.selectedWebView ? 0.5 : hoverTab == tab ? 0.2: 0.0))
-                            .frame(height: 50)
-                        
-                        HStack {
-                            if faviconLoadingStyle {
-                                WebImage(url: URL(string: "https://www.google.com/s2/favicons?domain=\(tab.url?.absoluteString)&sz=\(128)".replacingOccurrences(of: "https://www.google.com/s2/favicons?domain=Optional(", with: "https://www.google.com/s2/favicons?domain=").replacingOccurrences(of: ")&sz=", with: "&sz=").replacingOccurrences(of: "\"", with: ""))) { image in
-                                    image
-                                        .resizable()
-                                        .scaledToFit()
-                                        .frame(width: 25, height: 25)
-                                        .cornerRadius(50)
-                                        .padding(.leading, 5)
-                                    
-                                } placeholder: {
-                                    LoadingAnimations(size: 25, borderWidth: 5.0)
-                                        .padding(.leading, 5)
-                                }
-                                .onSuccess { image, data, cacheType in
-                                    // Success
-                                    // Note: Data exist only when queried from disk cache or network. Use `.queryMemoryData` if you really need data
-                                }
-                                .indicator(.activity) // Activity Indicator
-                                .transition(.fade(duration: 0.5)) // Fade Transition with duration
-                                .scaledToFit()
-                                
-                            } else {
-                                AsyncImage(url: URL(string: "https://www.google.com/s2/favicons?domain=\(tab.url?.absoluteString)&sz=\(128)".replacingOccurrences(of: "https://www.google.com/s2/favicons?domain=Optional(", with: "https://www.google.com/s2/favicons?domain=").replacingOccurrences(of: ")&sz=", with: "&sz=").replacingOccurrences(of: "\"", with: ""))) { image in
-                                    image
-                                        .resizable()
-                                        .scaledToFit()
-                                        .frame(width: 25, height: 25)
-                                        .cornerRadius(50)
-                                        .padding(.leading, 5)
-                                    
-                                } placeholder: {
-                                    LoadingAnimations(size: 25, borderWidth: 5.0)
-                                        .padding(.leading, 5)
-                                }
-                                
-                            }
-                            
-                            if tab.title == "" {
-                                Text(tab.url?.absoluteString ?? "Tab not found.")
-                                    .lineLimit(1)
-                                    .foregroundColor(Color.foregroundColor(forHex: UserDefaults.standard.string(forKey: "startColorHex") ?? "ffffff"))
-                                    .padding(.leading, 5)
-                                    .onReceive(timer) { _ in
-                                        reloadTitles.toggle()
-                                    }
-                            }
-                            else {
-                                Text(tab.title ?? "Tab not found.")
-                                    .lineLimit(1)
-                                    .foregroundColor(Color.foregroundColor(forHex: UserDefaults.standard.string(forKey: "startColorHex") ?? "ffffff"))
-                                    .padding(.leading, 5)
-                                    .onReceive(timer) { _ in
-                                        reloadTitles.toggle()
-                                    }
-                            }
-                            
-                            Spacer() // Pushes the delete button to the edge
-                            
-                            Button(action: {
-                                if let index = navigationState.webViews.firstIndex(of: tab) {
-                                    removeTab(at: index)
-                                }
-                            }) {
-                                if hoverTab == tab || navigationState.selectedWebView == tab {
-                                    ZStack {
-                                        Color(.white)
-                                            .opacity(hoverCloseTab == tab ? 0.3: 0.0)
-                                        
-                                        Image(systemName: "xmark")
-                                            .resizable()
-                                            .scaledToFit()
-                                            .frame(width: 15, height: 15)
-                                            .foregroundStyle(Color.white)
-                                            .opacity(hoverCloseTab == tab ? 1.0: 0.8)
-                                        
-                                    }.frame(width: 35, height: 35).cornerRadius(7).padding(.trailing, 10)
-                                        .hoverEffect(.lift)
-                                        .onHover(perform: { hovering in
-                                            if hovering {
-                                                hoverCloseTab = tab
-                                            }
-                                            else {
-                                                hoverCloseTab = WKWebView()
-                                            }
-                                        })
-                                    
-                                }
-                            }.keyboardShortcut("w", modifiers: .option)
-                        }
-                    }
-                    .contextMenu {
-                        Button {
-                            navigationState.createNewWebView(withRequest: URLRequest(url: URL(string: formatURL(from: tab.url?.absoluteString ?? ""))!))
-                        } label: {
-                            Label("Duplicate", systemImage: "plus.square.on.square")
-                        }
-                        
-                        Button {
-                            pinnedNavigationState.webViews.append(tab)
-                            
-                            if let index = navigationState.webViews.firstIndex(of: tab) {
-                                removeTab(at: index)
-                            }
-                        } label: {
-                            Label("Pin Tab", systemImage: "pin")
-                        }
-                        
-                        Button {
-                            favoritesNavigationState.webViews.append(tab)
-                            
-                            if let index = navigationState.webViews.firstIndex(of: tab) {
-                                removeTab(at: index)
-                            }
-                        } label: {
-                            Label("Favorite", systemImage: "star")
-                        }
-                        
-                        Button {
-                            if let index = navigationState.webViews.firstIndex(of: tab) {
-                                removeTab(at: index)
-                            }
-                        } label: {
-                            Label("Close Tab", systemImage: "xmark")
-                        }
-                        
-                    }
-                    .onHover(perform: { hovering in
-                        if hovering {
-                            hoverTab = tab
-                        }
-                        else {
-                            hoverTab = WKWebView()
-                        }
-                    })
-                    .onTapGesture {
-                        pinnedNavigationState.selectedWebView = nil
-                        pinnedNavigationState.currentURL = nil
-                        
-                        favoritesNavigationState.selectedWebView = nil
-                        favoritesNavigationState.currentURL = nil
-                        
-                        selectedTabLocation = "tabs"
-                        
-                        Task {
-                            await navigationState.selectedWebView = tab
-                            await navigationState.currentURL = tab.url
-                        }
-                        
-                        if let unwrappedURL = tab.url {
-                            searchInSidebar = unwrappedURL.absoluteString
-                        }
-                    }
-                    .onDrag {
-                        self.draggedTab = tab
-                        return NSItemProvider()
-                    }
-                    .onDrop(of: [.text], delegate: DropViewDelegate(destinationItem: tab, allTabs: $navigationState.webViews, draggedItem: $draggedTab))
+                    TodayTab(reloadTitles: $reloadTitles, tab: tab, hoverTab: $hoverTab, faviconLoadingStyle: $faviconLoadingStyle, searchInSidebar: $searchInSidebar, hoverCloseTab: $hoverCloseTab, selectedTabLocation: $selectedTabLocation, draggedTab: $draggedTab, navigationState: navigationState, pinnedNavigationState: pinnedNavigationState, favoritesNavigationState: favoritesNavigationState)
                 }
             }
             
@@ -710,7 +375,7 @@ struct Sidebar: View {
                             .resizable()
                             .scaledToFit()
                             .frame(width: 20, height: 20)
-                            .foregroundStyle(Color.white)
+                            .foregroundStyle(textColor)
                             .opacity(settingsButtonHover ? 1.0: 0.5)
                         
                     }.frame(width: 40, height: 40).cornerRadius(7)
@@ -800,7 +465,7 @@ struct Sidebar: View {
                                         .resizable()
                                         .scaledToFit()
                                         .frame(width: 20, height: 20)
-                                        .foregroundStyle(Color.white)
+                                        .foregroundStyle(textColor)
                                         .opacity(hoverSpace == space ? 1.0: 0.5)
                                     
                                 }.frame(width: 40, height: 40).cornerRadius(7)
